requirements.txt
google-genai
python-dotenv
colorama
SpeechRecognition
simpleaudio
elevenlabs
